server/core/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.1.0
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.6
redis==4.5.5
django-filter==23.2
django-environ==0.10.0
Pillow==10.0.0
requests==2.31.0
stripe==5.5.0
celery==5.3.1
django-celery-beat==2.5.0
gunicorn==21.2.0
whitenoise==6.5.0